+++ windows/flutter/generated_plugins.cmake
@@ -6,6 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   bonsoir_windows
   screen_retriever_windows
   tray_manager
+  url_launcher_windows
   window_manager
 )